feat(android): Document bootloader scripts
For SDK 10.1 and onwards, we now fetch the bootloaders from a
dedicated manifest using repo [1]

This manifest comes with a set of helper scripts, which can be ran to build all
required bootloader components (bl31, tee, tiboot3, tispl, u-boot).

Document how to use that instead of doing a lot of manual commands.

[1] https://git.ti.com/cgit/android/manifest/commit/?h=android15-release&id=6c17ffcc8396e56968b481cb961886ea3e634470
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>